Update Inventory_Management_System_Ver_0.1.py
Code has been rearchitectured, using an 'inventories manager' object to manage multiple inventories. Each inventory is also its own object, and can be used to add/remove assets from each inventory. 

Each inventory is now stored as a dictionary, with the 'key' being the name of the object and the 'value' being its quantity. 

The GUI is currently a basic command line, using numbers to select functions - this is subject to change to a GUI.