Create Inventory_Manager_GUI
To enhance ease of development, the GUI has been developed in this file. 

This will then be integrated with the 'inventory manager' file in the coming days/weeks to complete the system. 

The features of the GUI are as follows: 
- Add inventory button, remove inventory button, display inventory button, add item button, remove item button all added
- BAE Logo added 
- 'Inventory manager' title added
- Scroll box to show instant feedback of executed commands. This will show whether or not functions have been successfully executed, and will show inputted values. 
- Scroll box to show displayed inventories

These are yet to be organised into a proper GUI format - this is yet to be done.