Create Inventory_Manager_GUI
To enhance ease of development, the GUI has been developed in this file. 
This will then be integrated with the 'inventory manager' file in the coming days/weeks to complete the system. 
The features of the GUI are as follows: 
- Add inventory button, remove inventory button, display inventory button, add item button, remove item button all added - BAE Logo added 
- 'Inventory manager' title added 
- Scroll box to show instant feedback of executed commands. This will show whether or not functions have been successfully executed, and will show inputted values. 
- Scroll box to show displayed inventories These are yet to be organised into a proper GUI format - this is yet to be done.

Line 18: Child functions could not be accessed from the parent class. Using reference materials, such as Automate the boring stuff with Python identified that the super() function must be used on the class, enabling calling of the child functions from within the class. 

Line 58. Use of IntelliSense revealed syntax error. Originally, the function called was ‘scrolledtext.scrolledtext’, which was changed to ‘scrolledtext.Scrolledtext’ with a capital S to ensure the function was called correctly. Intellisense did not recognise the incorrect name for the function, and hence identified the error. 

Line 109 & 141. Original implementation of the code did not include this line. Each time the function was called, this caused ‘NameError: name ‘inventory_name/item_name’ is not defined. Peer debugging identified that the data in the entry must be predefined as a string with a variable length ‘StringVar()’, as now seen in the code. 

Lines 35, 39, 43, 47, 51. Originally, button commands were inputted as standard functions. Breakpoints were used to identify that standard functions are called as soon as the UI is created. Rubber ducking was used to identify that the functions should instead be made as ‘lambda’ functions, meaning they are only called when the button is clicked.