Create Command_Line_Inventory_Management_System
This is a version of the inventory management system with no UI (Uses command line to accept user input and output. This has been pulled directly from the 'Inventory_Management_System' code, but the GUI class has been removed. Once the GUI is completed, these will be integrated to form the overall system.